packages/backend/pyproject.toml
[project]
name = "translate-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",          # Web 框架，提供 API 路由和自动文档
    "uvicorn[standard]>=0.30.0", # ASGI 服务器，运行 FastAPI 应用
    "pydantic>=2.8.0",           # 数据校验和序列化
    "pydantic-settings>=2.4.0",  # 从 .env 文件自动读取配置
    "httpx>=0.27.0",             # 异步 HTTP 客户端（预留给百度翻译等 REST API）
    "redis[hiredis]>=5.0.0",     # 异步 Redis 客户端，缓存翻译结果
    "openai>=1.40.0",            # DeepSeek 的 API 兼容 OpenAI 协议，所以用这个 SDK 来调用
    "google-genai>=1.0.0",       # Google Gemini SDK，调用 Gemini 翻译引擎
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",             # 测试框架
    "pytest-asyncio>=0.23.0",    # pytest 异步测试支持
    "ruff>=0.5.0",               # 代码格式化 + Lint 工具
]